chore: remove read file feature
