+++ packages/docs/components/Playground.tsx
@@ -1,25 +1,13 @@
 import { DataStory } from '@data-story/ui'
-import { coreNodeProvider, Application, type Computer  } from '@data-story/core';
-import { ReadComputer } from './ReadFile';
-function App({ mode}: { mode?: 'js' | 'node' }) {
-  const originalRegister = coreNodeProvider.register;
+import { coreNodeProvider, Application } from '@data-story/core';
 
-  coreNodeProvider.register = (app: Application) => {
-    originalRegister(app);
-
-    const pwaComputer: Map<string, Computer> = new Map();
-    pwaComputer.set(ReadComputer.name, ReadComputer);
-
-    app.addComputers(pwaComputer);
-  }
-  // console.log(coreNodeProvider,'coreNodeProvider');
+export default ({ mode}: { mode?: 'js' | 'node' }) => {
   const app = new Application()
     .register(coreNodeProvider)
     .boot();
-  // console.log(app,'app');
 
   return (
-    <div style={{ height: '100vh', width: '100vw', overflow: 'hidden' }} data-cy="playground">
+    <div className="w-full" style={{ height: '100vh' }} data-cy="playground">
       <DataStory
         server={mode === 'node'
           ? { type: 'SOCKET', url: 'ws://localhost:3100' }
@@ -27,25 +15,4 @@ function App({ mode}: { mode?: 'js' | 'node' }) {
       />
     </div>
   );
-}
-
-export default App
-
-// import { DataStory } from '@data-story/ui'
-// import { coreNodeProvider, Application } from '@data-story/core';
-//
-// export default ({ mode}: { mode?: 'js' | 'node' }) => {
-//   const app = new Application()
-//     .register(coreNodeProvider)
-//     .boot();
-//
-//   return (
-//     <div className="w-full" style={{ height: '100vh' }} data-cy="playground">
-//       <DataStory
-//         server={mode === 'node'
-//           ? { type: 'SOCKET', url: 'ws://localhost:3100' }
-//           : { type: 'JS', app }}
-//       />
-//     </div>
-//   );
-// };
+};
