+++ packages/pwa/src/App.tsx
@@ -1,22 +1,21 @@
 import { DataStory } from '@data-story/ui'
 import { coreNodeProvider, Application, type Computer  } from '@data-story/core';
-import { ReadComputer } from "./ReadFile.ts";
+import { ReadComputer } from './ReadFile';
 function App({ mode}: { mode?: 'js' | 'node' }) {
   const originalRegister = coreNodeProvider.register;
 
   coreNodeProvider.register = (app: Application) => {
-   originalRegister(app);
+    originalRegister(app);
 
-   const pwaComputer: Map<string, Computer> = new Map();
-   pwaComputer.set(ReadComputer.name, ReadComputer);
+    const pwaComputer: Map<string, Computer> = new Map();
+    pwaComputer.set(ReadComputer.name, ReadComputer);
 
-   app.addComputers(pwaComputer);
+    app.addComputers(pwaComputer);
   }
-  // console.log(coreNodeProvider,'coreNodeProvider');
+
   const app = new Application()
     .register(coreNodeProvider)
     .boot();
-  // console.log(app,'app');
 
   return (
     <div style={{ height: '100vh', width: '100vw', overflow: 'hidden' }} data-cy="playground">
