+++ packages/pwa/src/ReadFile.ts
@@ -1,14 +1,18 @@
-import { ComputerConfig, ComputerFactory } from "@data-story/core";
+import { ComputerConfig, ComputerFactory } from '@data-story/core';
+
+interface FileSystemFileHandle {
+  getFile: () => Promise<File>;
+}
 
 export const ReadFile: ComputerConfig = {
   name: 'ReadFile',
-  inputs: [],
-  outputs: [],
+  inputs: ['input'],
+  outputs: ['path', 'content'],
   params: [
     {
-      name: 'path',
-      label: 'Path',
-      help: 'Path to file',
+      name: 'selectFile',
+      label: 'select csv/json file',
+      help: '',
       inputMode: {
         type: 'SelectFile',
         path: '',
@@ -18,41 +22,28 @@ export const ReadFile: ComputerConfig = {
     }
   ],
 
-  // eslint-disable-next-line no-empty-pattern
-  async *run({}) {},
+  canRun({input}) {
+    return input.haveAllItemsAtInput('input');
+  },
+
+  async *run({output, params}) {
+
+    const fileHandle = params['selectFile'] as unknown as FileSystemFileHandle;
+    const file = await fileHandle?.getFile?.();
+    const content = await file?.text();
+    const path = (fileHandle as any).name;
+
+    output.pushTo('content', [{
+      content: content
+    }]);
+    output.pushTo('path', [{
+      path: path
+    }]);
+
+    yield;
+  },
 };
 
 export const ReadComputer = new ComputerFactory().get(ReadFile);
-
-interface FileSystemFileHandle {
-  getFile: () => Promise<File>;
-}
-
-export const readFileFn = async() => {
-  const handleFileRead = async(fileHandle: FileSystemFileHandle) => {
-    const file = await fileHandle.getFile();
-    return await file.text();
-  };
-
-  try {
-    // show file picker
-    const [fileHandle] = await window.showOpenFilePicker!({
-      types: [
-        {
-          description: 'Text Files',
-          accept: {
-            'text/csv': ['.csv'],
-            'application/json': ['.json'],
-          },
-        },
-      ],
-    });
-
-    return await handleFileRead(fileHandle as unknown as FileSystemFileHandle);
-  } catch(error) {
-    console.error('File selection was cancelled or failed', error);
-  }
-};
-
